Moving some files and fixing some compliance
