+++ Library_Engine/Query/Libraries.cs
@@ -39,9 +39,9 @@ namespace BH.Engine.Library
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Gets the content of the Datasets from the library")]
-        [Input("libraryName", "The name of the Dataset(s) to extract")]
-        [Output("libraryData", "The data from the Dataset(s)")]
+        [Description("Gets the content of the Datasets from the library.")]
+        [Input("libraryName", "The name of the Dataset(s) to extract.")]
+        [Output("libraryData", "The data from the Dataset(s).")]
         public static List<IBHoMObject> Library(string libraryName)
         {
             return Datasets(libraryName).SelectMany(x => x.Data).ToList();
